requirements.txt
torch
torchvision
transformers
hdbscan
scikit-learn
numpy
Pillow
tqdm
